Seal the standalong LexerParserException hierarchy
